api/requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.9.2
cryptography==45.0.6
daphne==4.2.1
decorator==5.2.1
distlib==0.3.9
dj-rest-auth==7.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-auth==0.9.5
django-stubs==5.2.1
django-stubs-ext==5.2.1
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
executing==2.2.0
factory_boy==3.3.3
Faker==37.4.0
filelock==3.18.0
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack==1.1.1
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
priority==2.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
ruff==0.12.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
wsproto==1.2.0
zope.interface==7.2
